entity 대량 수정 및 admincartassign 구현

+++ src/main/java/com/springboot/bizconnect/domain/cart/dto/detail/CompanyProductDetailResponseDto.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CompanyProductDetailResponseDto {
-	private int productNo;
+	private Long productNo;
 	private String unitName;
 	private String categoryName;
 	private String manufacturerName;
