+++ src/main/java/com/springboot/bizconnect/domain/product/dto/detail/ProductDetailResponseDto.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDetailResponseDto {
-	private int productNo;
+	private Long productNo;
 	private String unitName;
 	private String categoryName;
 	private String manufacturerName;
