+++ src/main/java/com/springboot/bizconnect/entity/Message.java
@@ -29,7 +29,7 @@ public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "message_no", nullable = false)
-    private Integer messageNo;
+    private Long messageNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_no", nullable = false)
