+++ src/main/java/com/springboot/bizconnect/entity/Notice.java
@@ -26,7 +26,7 @@ public class Notice extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "notice_no", nullable = false)
-    private Integer noticeNo;
+    private Long noticeNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_no", nullable = false)
